Update param_array tests to actually run on GPU
For the param_array tests, I was catching the RuntimeError that is
raised if you pass code into code_isotropic and are on the GPU. So a lot
of these tests were effectively being skipped (or rather testing the
error was raised instead of testing the actual energy calculation). To
fix this, I added another test to run on the GPU that does not pass any
code into code_isotropic. I changed the original test to only run on the
CPU so that the full test is run.