Refactor how patch classes inherit from base class
There were some things in the base class that were specific to the
singlet class, e.g., the property `code`. Essentially, we were trying to
pigeonhole things into the base class that didn't really belong there.
This commit changes that, so that only functions that are truly shared
between the derived classes are in the base class.