Fix TypeConverter.__call__'s behavior with RequiredArg
Allows RequiredArg to get through the validation which is necessary
when querying nonexistent keys. Since we changed __getitem__ for
TypeParameterDict to correctly create querried keys, the default must be
settable for all keys even when unspecified. This should not result in
any errors unless, the user explicitly tries to set a value with
RequiredArg in it.