Fix flake8 errors
I use a lot of noqa here. Most of them are in classes that are soon to
be ported to v3 or removed from the codebase. I also noqa'ed the
entire docstring for the Harmonic class because I couldn't get the
return types to display properly using line continuing backslashes and I
did not want to waste any more time tinkering with it.