Move hpmc walls to separate file + remove classes
This commit has a lot of deletions. I totally removed the frenkel-ladd,
composite, and callback classes.

Since Frenkel-Ladd was essentially just a harmonic field plus a remove
drift updater, we are not keeping that class and will instead make a
tutorial for the examples repo.

Composite fields will get totally restructured in the future, so the
existing class is not needed, hence the removal.

The callback functions are also no longer needed since we have the JIT
external fields.

The wall class is also going to be significantly re-worked, and will be
more similar to the MD walls structure; I therefore removed that class
too.