Refactor ExternalFieldHarmonic.getEnergy()
Instead of passing those boolean agruments down the line, this function
now just calls the corresponding calcE_X functions (X = trans/rot) and
returns a std::pair, one item for the translational energy and the other
for the rotational energy. This commit also changes the corresponding
functions in the python class.