Fix MPI logic in computeTailCorrection()
Since the energy and pressure corrections are global values, these only
need to be done on the root rank, and the values do not need to be
reduced on all ranks. The only thing that needs to be communicated
across all ranks is the number of particles of each type, which is used
to compute the number densities, which go into the tail correction
calculations.