Change name of spheropolyhedron templated wall class
In the python wall code, the c++ class name is constructed by
concatenating "Wall" + the name of the integrator class. For some
reason, the python hpmc integrator associated with the c++ class
ShapeSpheropolyhedron is called ConvexSpheropolyhedron. This change
makes it so that the templated wall class has the same name that is
expected in the python code.