Add more robust wall-shape validation checks
The initial implementation only checked that the wall types were
compatible with the hpmc integrator when walls were set using the
setter. Users could still append walls to the wall list exposed in
python, leading to behavior inconsistent with what would happen using
the walls setter. This commit subclasses the _WallsMetaList for HPMC
walls and overwrites the appropriate functions to check for validity.

I also added tests to test this behavior.