Fix bugs in random search SDF for patchy particles
My initial implementation was scaling r_ij incorrectly, and there was
also some errors in the logic for selecting which bin to sample. The
problem was if min_bin was 0, then the random number was selected in the
range (0, -1) = (0, 2^32), so I was getting a lot of garbage in the SDF
which of course gave me garbage pressure results. This commit fixes all
of those bugs, and I can reproduce the pressure in an NVT simulation of
patchy triangular particles.