feat: Add BinaryOptimizer to solvers
The motivation for this feature is to allow for an optimization protocol
that tests a variety of values without getting stuck in a local extrema.

The solver adds a bucket optimization scheme where the domain is divided
into buckets and each bucket is tested for is y value. The best bucket
is chosen after all are tested, and if the desired number of rounds is
met the bucket center is the final optimization. If the number of rounds
is greater than the process repeats with bin extent serving as the new
domain.