Implement brute-force search in per-pair SDF
This commit implements a brute-force overlap search in the SDF
calculation for enthalpic systems. My previous attempts to use a random
search algorithm with the per-pair overlap check (no commits made)
yielded calculated pressures that were very far off of the reference
pressures, so this commit is attempt to do the per-pair calculation with
the brute force method (i.e., start at bin 0, and evaluate overlaps at
sequentially increasing bins until the first one is found).

Unfortunately, I am still unable to reproduce the correct pressure with
this method. I have tried several different normalizations of the SDF
(number of histogam entries, total number of pairs), but to no avail.

This commit is here for reference to see what I tried, but will be
reverted. Following commit(s) will re-implement the per-particle SDF
algorithm with the brute force approach taken here. It seems like we
should be able to take advantage of all the per-pair overlaps that we
calculate to generate better statistics, but we are already getting very
accurate pressure measurements with the per-particle method, so I will
likely hold off on digging much deeper here and continue using the
per-particle version.