Implement per-particle SDF with brute-force search
This commit reverts the per-pair SDF algorithm back to the per-particle
version. It is also the first version of the code that uses the
brute-force first overlap search for the SDF that accounts for enthalpic
interactions. With this version of the code, I am able to reproduce the
pressure of a system of patchy hexagons to ~2 decimals.