Implement expansion moves for SDF
The SDF method that calculates the pressure in systems with
discontinuous potentials (e.g., hard particles) uses volume
perturbations. As originally implemented for convex hard particles, only
negative volume perturbations are implemnted since they are the only
volume prturbations that contribute to the pressure in those systems.
However, some systems, e.g., systems with discontinuous pair potentials
or convex hard particles, will have meaningful contributions to the
pressure from epansive volume perturbations.

This commit implements the expansion moves in addition to the
compression moves. There is a lot of code duplication because I was
first testing whether or not we can get reasonable answers before making
the code more elegant. So now we keep track of two scale distribution
functions: one for expansive and one for compressive volume
perturbations. The pressure that is returned is a weighted average of
the pressure calculated from expansion and compression moves, weighted
by the number of entries in each respective histogram. However, there is
nothing from the theory that says this should be a weighted average, and
instead should be an unweighted average; I used the weighted average
because that gave me pressures that were closer to the "ground truth"
pressure in my tests. Hence, use this code at your own risk. Further
theoretical developments will work address this inconsistency between
theory and simulation.