Correctly compute pressure with expansion moves
The correct way to compute pressure when doing both expansion and
compression moves is to simply sum their contributions to the excess
pressure. This algorithm gives the expected result for hard convex
particles where isotropic expansions cannot result in overlaps, so the
contribution to the (excess) pressure is zero for those moves. I am able
to get pressure/density agreement with constant-volume and
constant-pressure simulations for a simple concave paticle model.