Add expansion moves to ComputeSDF
This commit adds expansive volume perturbations to the SDF compute.
These perturbations are necessary for nonconvex hard shapes or patchy
"nonmonotonic" patchy interactions (monotonic is probably not the right
word here, but we have not been able to come up with a better word; see
https://github.com/glotzerlab/hoomd-blue/pull/1356#discussion_r943488535
for some discussion on this topic).

To check if expansion moves are need when the compute is triggered, I
added a requiresExpansiveSDF() method to IntegratorHPMCMono and the
different shape classes. The integrator's method returns `true` if the
method for any of the shapes in the system return `true`. The relevant
member variable of the integrator is set during prepRun(). Whenever the
linear search patch is taken, expansion moves are performed. The linear
search path is always taken if either the integrator's
requiresExpansiveSDF method returns true or if there are patchy
interactions in the system, and therefore systems with patches always do
expansion moves.