Fix a bug in linear search SDF calculation
I was getting the wrong pressures because of a bug I introduced when
adding the expansion moves. The bug was that I was not carrying over the
minimum distance to overlap from the i-j1 pair to i-j2. Hence, the SDF I
was calculating was too high at larger scale values than it should have
been, and hence the pressure was being underestimated compared to the
true value. This commit fixes that bug by comparing the current bin
(corresponding to the scale value) to the minimum bin where an overlap
has been found and skipping the overlap calculation if the current bin
is greater than the minimum bin.