feat: Allow for intra operation list dependencies.
We will no longer call `~._attach` twice in this case. This introduces
a new internal API `~._attach_hook` and `~._detach_hook` which provides
customization upon attaching and detaching while the calling methods
keep track of the number of times they have been called as a way to know
the number of consumers that an object has.