Update rigid bodies both before and after communication.
This ensures that constituent particles are migrated to the proper
domain before forces are calculated. Without the call before, the
rigid body centers are moved but the constituents left at the previous
position. In this way, a constituent that should migrate to a new rank
might be left behind - outside the ghost particle window for a neighbor
rank. This leads to i->j force calculations without the corresponding
j->i calculation on a neighbor rank.

By updating bodies before communicating, the constituent particles
that move outside the local domain are communicated correctly. The
update bodies call after communication is required to ensure that
constituents are updated properly on ranks where both the constituent
AND the central particle are ghosts.