Correct the computation of the body ghost width.
Constituent particles need no extra ghost width. Only those within
the normal r_cut + r_buff range of a boundary need to be communicated.

Only the rigid body central particles *may* need an extra ghost
width. Specifically, the worse case scenario is when a constituent
ghost is exactly r_ghost_width away from the boundary. To compute
the correct position, the central particle must also be a ghost,
so the minimum ghost width for a central particle is
max[d_i + r_ghost_i] over all particles in the body. This
sets a minimum. If the ghost width for the body central particle is
already greater than this, it need not be increased.