Remove neighborlist error based on rigid body diameters.
This error was incorrectly determined. The minimum image convention
is applied only between individual particles, not bodies. If the user
chooses to make a rigid body as large as the box, the simulation may
be ill-stated but should run. There isn't much HOOMD can do to protect
users from the egregious error of defining a rigid body larger than the
simulation box.