Add body ghost width for constituents.
In previous commits on this branch, I incorrectly assumed that the
body ghost width for constituents should be 0. My test case involved
bodies where the r_cut for constituents was large. The body net force
computation requires that the rank hosting the central particle must
be able to access all constituents as ghosts so that it can sum the
net force and torque. Adjust the body ghost width computation to
account for this.