Correct additional corner cases.
In the previous commits on this branch, I assumed that the rigid body
did not need to be complete. This assumption was incorrect, as the way
that ForceComposite indexes particles from the molecule in reference to
the body assumes that the body is complete. Expand the constituent
ghost width to include this.

With the additional ghost width, there are now many extra constituent
ghost particles that cannot access their central particle. However, they
do not need to because they are outside the pair force interaction range
of the domain boundary. Therefore, we need to allow for non-local
central particles. By construction of the body ghost widths, any non-local
central particle should not be needed. However, there is no straightforward
way to check for this error condition as updateCompositeParticles does not
have knowledge of the interaction ghost widths to verify that every
constituent ghost within the interaction width has a central particle.