Update particle counting for DOF removal.
Count only the central and free particles in the group and the system
when considering how many DOF to remove due to periodic boundary conditions.

I had to remove the `const` on all the ParticleGroup methods because
none of them are really const - they all modify the internal data
buffers of the ParticleGroup even if they do not change its contents.