fix: Write first frame in GSDDequeWriter
While this is unexpected, it is required to keep performance high as
until the first write for a file occurs, all data is recorded and sorted
in a GSDDumpWriter::GSDFrame object. By writing the first frame, we
prevent this. However, we then don't know if the first frame should be
included in the actual analysis.