Use sybil to test ConstantPressure code examples.
Classes from methods.py are imported into __init__.py. Sybil executes
all code in a single .py file in the same namespace. This commit duplicates
the setup code in __init__.py and methods.py. The code shown in __init__.py
is shown to the user in the rendered HTML output. The code in methods.py
is needed to prepare the namespace for the examples to run.