Use proper gather to sort forces, torques, and virials
I added a vec6 Struct to hold the virials since you cannot create
Vector-s of arrays. I also updated md/pytest/test_custom_force.py to set
the values of the local force arrays to be functions of the particle tag
to test that the values are returned in the proper order.