Deprecate depletant fugacity.
Also switch to stacklevel=1. warnings.warn only prints 1 (ONE) line
stating the source of the warning. Depending on how the user attaches,
the number of levels between user code and this function vary.

stacklevel=2 may be appropriate in other places in HOOMD (e.g. properties
that are typically accessed directly by the user).