Add getSortedLocalTags helper function
Every get*Python() function in ForceCompute does the same thing, so this
commit adds a helper function to call in each of those functions. I
implemented the function and use it in getEnergiesPython() here, and it
compiles and passes tests. Next commit will update the other
get*Python() functions to use getSortedLocalTags().