Add VectorVariant class
Adds a VectorVariant class.
Adds VectorVariantBox as a subclass of VectorVariant.
Implements VectorVariantBoxConstant and VectorVariantBoxLinear.
Moves variant from a module to a package.

The code compiles and I am able to get the expected boxes as a function
of time for both VectorVariantBoxConstant and VectorVariantBoxLinear
objects.

More to come before I open a PR.