Move scalar variants to new module
Moves the scalar variant classes to a separate module, but imports them
in hoomd/variant/__init__.py so that they still live in the
hoomd.variant namespace. I did this as a workaround for python's import
behavior since I kept getting circular import errors when keeping the
scalar variant classes in the package's __init__.py file.