Update failing test and mark with pytest.validate
`test_inverse_volume_slow_compress` was failing on MPI because the box
was too small for domain decomposition, so this commit uses a larger
system.

That test also takes ~4 seconds to run in serial, so mark it with
pytest.validate since we do not gain anything from running that test
on every commit.