Start adding tests for external potentials.
I did not get very far because I ran into errors in my new test that I
could not figure out how to fix.
I wrote a test that tests attaching external potentials to the
simulation.
However, an exception is raised when HOOMD tries to attach the linear
external potential.
The error message I get is
```
TypeError: (): incompatible function arguments. The following argument types are supported:
    1. (arg0: hoomd.hpmc._hpmc.ExternalPotentialLinear, arg1: hoomd._hoomd.vec3_double) -> None

Invoked with: <hoomd.hpmc._hpmc.ExternalPotentialLinear object at 0x10706f3f0>, (0.0, 0.0, 0.0)
```

The 2nd argument in the "invoked with:" list is a `_HOOMDTuple` and not
a `vec3_double`. I'm not sure if that's the cause of the error, but it's
my best guess.