Check that linker is active now and after virtual move
Must totally abort the cluster move if any of the linker particles start
or end up in the inactive region.

Perhaps an optimization opportunity: once a linkee has been definitely
added to the cluster, check its current and trial position to see if it
starts or ends up in the inactive region. That would give you a much
earlier early exit (compared to what we do now) since you wouldn't have
to continue checking the rest of the current or following linkers'
neighbors for overlaps/pair energies.